third_party/blink/web_tests/svg/hixie/error/001-expected.txt

CONSOLE ERROR: Error: <circle> attribute r: Expected length, "r".
